refactor(jaml): move flow parser to jaml module
