+++ jina/drivers/__init__.py
@@ -19,7 +19,7 @@ from ..executors import BaseExecutor
 from ..executors.compound import CompoundExecutor
 from ..executors.decorators import wrap_func
 from ..helper import convert_tuple_to_list
-from ..jaml import JAML
+from ..jaml import JAMLCompatible
 
 if False:
     # fix type-hint complain for sphinx and flake
@@ -71,8 +71,7 @@ def store_init_kwargs(func: Callable) -> Callable:
 
 
 class QuerySetReader:
-    """
-    :class:`QuerySetReader` allows a driver to read arguments from the protobuf message. This allows a
+    """:class:`QuerySetReader` allows a driver to read arguments from the protobuf message. This allows a
     driver to override its behavior based on the message it receives. Extremely useful in production, for example,
     get ``top_k`` results, doing pagination, filtering.
 
@@ -92,16 +91,19 @@ class QuerySetReader:
         - the ``priority`` in the proto should be strictly greater than the driver's priority (by default is 0)
         - the field name must exist in proto's ``parameters``
 
+    .. warning::
+        For the sake of cooperative multiple inheritance, do NOT implement :meth:`__init__` for this class
+
     """
 
     def _get_parameter(self, key: str, default: Any):
         if getattr(self, 'queryset', None):
             for q in self.queryset:
                 if (
-                    not q.disabled
-                    and self.__class__.__name__ == q.name
-                    and q.priority > self._priority
-                    and key in q.parameters
+                        not q.disabled
+                        and self.__class__.__name__ == q.name
+                        and q.priority > self._priority
+                        and key in q.parameters
                 ):
                     ret = q.parameters[key]
                     return dict(ret) if isinstance(ret, Struct) else ret
@@ -117,7 +119,7 @@ class QuerySetReader:
         raise AttributeError
 
 
-class DriverType(type):
+class DriverType(type(JAMLCompatible), type):
     def __new__(cls, *args, **kwargs):
         _cls = super().__new__(cls, *args, **kwargs)
         return cls.register_class(_cls)
@@ -131,11 +133,10 @@ class DriverType(type):
 
             reg_cls_set.add(cls.__name__)
             setattr(cls, '_registered_class', reg_cls_set)
-        JAML.register(cls)
         return cls
 
 
-class BaseDriver(metaclass=DriverType):
+class BaseDriver(JAMLCompatible, metaclass=DriverType):
     """A :class:`BaseDriver` is a logic unit above the :class:`jina.peapods.pea.BasePea`.
     It reads the protobuf message, extracts/modifies the required information and then return
     the message back to :class:`jina.peapods.pea.BasePea`.
@@ -215,24 +216,6 @@ class BaseDriver(metaclass=DriverType):
             r['with'] = a
         return r
 
-    @classmethod
-    def to_yaml(cls, representer, data):
-        """Required by :mod:`pyyaml` """
-        tmp = data._dump_instance_to_yaml(data)
-        return representer.represent_mapping('!' + cls.__name__, tmp)
-
-    @classmethod
-    def from_yaml(cls, constructor, node):
-        """Required by :mod:`pyyaml` """
-        return cls._get_instance_from_yaml(constructor, node)
-
-    @classmethod
-    def _get_instance_from_yaml(cls, constructor, node):
-        data = constructor.construct_mapping(node, deep=True)
-
-        obj = cls(**data.get('with', {}))
-        return obj
-
     def __eq__(self, other):
         return self.__class__ == other.__class__
 
@@ -257,12 +240,12 @@ class BaseRecursiveDriver(BaseDriver):
     # TODO(Han): probably want to publicize this, as it is not obvious for driver
     #  developer which one should be inherited
     def _apply_all(
-        self,
-        docs: 'DocumentSet',
-        context_doc: 'Document',
-        field: str,
-        *args,
-        **kwargs,
+            self,
+            docs: 'DocumentSet',
+            context_doc: 'Document',
+            field: str,
+            *args,
+            **kwargs,
     ) -> None:
         """Apply function works on a list of docs, modify the docs in-place
 
@@ -341,8 +324,8 @@ class BaseExecutableDriver(BaseRecursiveDriver):
     def exec_fn(self) -> Callable:
         """the function of :func:`jina.executors.BaseExecutor` to call """
         if (
-            not self.msg.is_error
-            or self.pea.args.skip_on_error < SkipOnErrorType.EXECUTOR
+                not self.msg.is_error
+                or self.pea.args.skip_on_error < SkipOnErrorType.EXECUTOR
         ):
             return self._exec_fn
         else:
@@ -357,7 +340,7 @@ class BaseExecutableDriver(BaseRecursiveDriver):
             else:
                 for c in executor.components:
                     if any(
-                        t.__name__ == self._executor_name for t in type.mro(c.__class__)
+                            t.__name__ == self._executor_name for t in type.mro(c.__class__)
                     ):
                         self._exec = c
                         break
