+++ jina/jaml/parsers/flow/v1.py
@@ -46,7 +46,7 @@ class V1Parser(VersionedYAMLParser):
     """
     version = '1'  # the version number this parser designed for
 
-    def parse(self, data: Dict) -> 'Flow':
+    def parse(self, cls: type, data: Dict) -> 'Flow':
         """Return the Flow YAML parser given the syntax version number
 
         :param data: flow yaml file loaded as python dict
@@ -58,7 +58,7 @@ class V1Parser(VersionedYAMLParser):
         # maybe there are some hanging kwargs in "parameters"
         tmp_a = (expand_env_var(v) for v in a)
         tmp_p = {kk: expand_env_var(vv) for kk, vv in {**k, **p}.items()}
-        obj = Flow(*tmp_a, env=envs, **tmp_p)
+        obj = cls(*tmp_a, env=envs, **tmp_p)
 
         pp = data.get('pods', [])
         for pods in pp:
