+++ tests/unit/flow/test_flow_yaml_parser.py
@@ -4,10 +4,13 @@ import numpy as np
 import pytest
 
 from jina import Flow, AsyncFlow
+from jina.enums import FlowOptimizeLevel
 from jina.excepts import BadFlowYAMLVersion
 from jina.flow import BaseFlow
-from jina.jaml.parsers import get_supported_versions
 from jina.jaml import JAML
+from jina.jaml.parsers import get_supported_versions
+from jina.parser import set_flow_parser
+from tests import rm_files
 
 cur_dir = Path(__file__).parent
 
@@ -56,3 +59,36 @@ def test_load_dump_load(tmpdir):
     f1.save_config(str(Path(tmpdir) / 'a0.yml'))
     f2 = Flow.load_config('yaml/flow-v1.0-syntax.yml')
     f2.save_config(str(Path(tmpdir) / 'a1.yml'))
+
+
+def test_load_flow_with_port():
+    f = Flow.load_config('yaml/test-flow-port.yml')
+    with f:
+        assert f.port_expose == 12345
+
+
+def test_load_flow_from_cli():
+    a = set_flow_parser().parse_args(['--uses', 'yaml/test-flow-port.yml'])
+    f = Flow.load_config(a.uses)
+    with f:
+        assert f.port_expose == 12345
+
+
+def test_load_flow_from_yaml():
+    with open(cur_dir.parent / 'yaml' / 'test-flow.yml') as fp:
+        a = Flow.load_config(fp)
+        with open(cur_dir.parent / 'yaml' / 'swarm-out.yml', 'w') as fp, a:
+            a.to_swarm_yaml(fp)
+        rm_files([str(cur_dir.parent / 'yaml' / 'swarm-out.yml')])
+
+
+def test_flow_yaml_dump():
+    f = Flow(logserver_config=str(cur_dir.parent / 'yaml' / 'test-server-config.yml'),
+             optimize_level=FlowOptimizeLevel.IGNORE_GATEWAY,
+             no_gateway=True)
+    f.save_config('test1.yml')
+
+    fl = Flow.load_config('test1.yml')
+    assert f.args.logserver_config == fl.args.logserver_config
+    assert f.args.optimize_level == fl.args.optimize_level
+    rm_files(['test1.yml'])
